Added device event subscription

+++ src/particle.android.ts
@@ -3,6 +3,8 @@ import * as utils from "tns-core-modules/utils/utils";
 
 // keep this baby active while logged in as it holds state (our devices)
 let worker;
+let eventWorker;
+var cbArr: any = undefined;
 
 declare const io: any;
 const ParticleCloudSDK = io.particle.android.sdk.cloud.ParticleCloudSDK;
@@ -31,10 +33,25 @@ export class Particle implements TNSParticleAPI {
     });
   }
 
+  public loginWithToken(token: string): void {
+    ParticleCloudSDK.getCloud().setAccessToken(token);
+    if (global["TNS_WEBPACK"]) {
+      const WorkerScript = require("nativescript-worker-loader!./particle-worker.js");
+      worker = new WorkerScript();
+    } else {
+      worker = new Worker("./particle-worker.js");
+    }
+  }
+ 
+  public setOAuthConfig(id: string, secret: string): void {
+
+  }
+
   public logout(): void {
     // no need for a worker here because there are no network calls involved
     ParticleCloudSDK.getCloud().logOut();
-    worker.terminate();
+    if (worker) worker.terminate();
+    if (eventWorker) eventWorker.terminate();
   }
 
   public listDevices(): Promise<Array<TNSParticleDevice>> {
@@ -42,6 +59,7 @@ export class Particle implements TNSParticleAPI {
       worker.postMessage({
         action: "listDevices"
       });
+      
 
       worker.onmessage = msg => {
         if (msg.data.success) {
@@ -50,7 +68,23 @@ export class Particle implements TNSParticleAPI {
           devices.map(device => {
             device.callFunction = (name: string, args): Promise<number> => this.callFunction(device.id, name, args);
             device.getVariable = (name: string): Promise<any> => this.getVariable(device.id, name);
+            device.subscribe = (name: string, eventHandler: any): void => this.subscribe(device.id, name, eventHandler);
+            device.unsubscribe = (): void => this.unsubscribe(device.id);
           });
+
+          // start event subscription worker and get device list
+          if (!eventWorker) {
+            if (global["TNS_WEBPACK"]) {
+              const EventWorkerScript = require("nativescript-worker-loader!./particle-event-worker.js");
+              eventWorker = new EventWorkerScript();
+            } else {
+              eventWorker = new Worker("./particle-event-worker.js");
+            }
+          }
+          eventWorker.postMessage({
+            action: "listDevices"
+          });
+
           resolve(devices);
         } else {
           reject(msg.data.error);
@@ -86,5 +120,56 @@ export class Particle implements TNSParticleAPI {
 
       worker.onmessage = msg => msg.data.success ? resolve(msg.data.result) : reject(msg.data.error);
     });
+  }
+
+  private unsubscribe(deviceId: string): void {
+    eventWorker.postMessage({
+      action: "unsubscribe",
+      options: { 
+        deviceId}
+    });
+    cbArr = undefined;
+  }
+
+  private subscribe(deviceId: string, name: string, eventHandler: any): void {
+    if (cbArr == undefined) { 
+      cbArr = {}; 
+    }
+    cbArr[name] = eventHandler;
+    console.dir(cbArr);
+    
+    eventWorker.postMessage({
+      action: "subscribe",
+      options: { 
+        deviceId,
+        name
+      }
+    });
+
+    eventWorker.onmessage = (msg) => {
+      if (msg.data.success) {
+        const d = msg.data.data;
+        // console.log(`${d.name}: ${d.data}`);
+        var cb : (any) => any;
+        cb = cbArr[d.name];
+        if (cb) cb(d.data);
+      }
+    }
+  }
+
+  public isAuthenticated(): Boolean {
+    return ParticleCloudSDK.getCloud().isLoggedIn();
+}
+  
+  public accessToken(): string {
+    return ParticleCloudSDK.getCloud().getAccessToken();
+  }
+
+  public startDeviceSetupWizard(cb:any): void {
+    // stub for startDeviceSetupWizard
+  }
+
+  public getDeviceSetupCustomizer(): any {
+    // stub for getDeviceSetupCustomizer
   }
 }
